requirements.txt
livekit>=1.0.0
"numpy<2"
faiss-cpu
sentence-transformers
livekit-agents 
livekit-plugins-google==1.3.3
flask 
flask-cors 
python-dotenv 
aiohttp